requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
colorama>=0.4.6
adafruit-circuitpython-max31856==0.11.6
adafruit-circuitpython-rgb-display==3.11.3
appdirs==1.4.4
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
pyserial==3.5
